run systemctl stop netclient only if daemon installed

+++ daemon/systemd.go
@@ -76,10 +76,12 @@ func RestartSystemD() {
 
 // CleanupLinux - cleans up neclient configs
 func CleanupLinux() {
-	if _, err := ncutils.RunCmd("systemctl stop netclient", false); err != nil {
-		logger.Log(0, "failed to stop netclient service", err.Error())
+	if config.Netclient.DaemonInstalled {
+		if _, err := ncutils.RunCmd("systemctl stop netclient", false); err != nil {
+			logger.Log(0, "failed to stop netclient service", err.Error())
+		}
+		RemoveSystemDServices()
 	}
-	RemoveSystemDServices()
 	if err := os.RemoveAll(config.GetNetclientPath()); err != nil {
 		logger.Log(1, "Removing netclient configs: ", err.Error())
 	}
